refactor(sql): move AUTO datetime detection to common-expression and simplify eval functions
Move auto-detect datetime parsing functions (int64_to_timestamp,
parse_epoch_str, auto_detect_*, parse_*_with_auto, etc.) from
databend-functions-scalar-datetime into a new module at
databend-common-expression::utils::auto_detect_datetime.

This fixes the inverted dependency where databend-common-formats
depended on databend-functions-scalar-datetime. The formats crate
now imports these utilities directly from common-expression.

Also extract the 4-layer fallback logic (ISO -> epoch/auto -> dtparse)
from eval_string_to_timestamp and eval_string_to_date into standalone
parse_string_to_timestamp and parse_string_to_date functions, making
the vectorized eval closures concise.

Fix setting descriptions: mention enable_auto_detect_datetime_format
interaction in enable_strict_datetime_parser, and add missing Unix
date format to enable_auto_detect_datetime_format description.
