+++ src/query/expression/src/utils/mod.rs
@@ -14,6 +14,7 @@
 
 pub mod arithmetics_type;
 pub mod arrow;
+pub mod auto_detect_datetime;
 pub mod bitmap;
 pub mod block_debug;
 pub mod block_thresholds;
