+++ src/query/formats/src/field_decoder/common.rs
@@ -19,15 +19,15 @@ use databend_common_column::types::timestamp_tz;
 use databend_common_exception::ErrorCode;
 use databend_common_exception::Result;
 use databend_common_expression::types::timestamp::clamp_timestamp;
+use databend_common_expression::utils::auto_detect_datetime::auto_detect_timestamp;
+use databend_common_expression::utils::auto_detect_datetime::int64_to_timestamp;
+use databend_common_expression::utils::auto_detect_datetime::parse_date_with_auto;
+use databend_common_expression::utils::auto_detect_datetime::parse_epoch_str;
+use databend_common_expression::utils::auto_detect_datetime::parse_timestamp_tz_with_auto;
 use databend_common_io::cursor_ext::BufferReadDateTimeExt;
 use databend_common_io::cursor_ext::DateTimeResType;
 use databend_common_io::cursor_ext::ReadBytesExt;
 use databend_common_io::cursor_ext::read_num_text_exact;
-use databend_functions_scalar_datetime::datetime::auto_detect_timestamp;
-use databend_functions_scalar_datetime::datetime::int64_to_timestamp;
-use databend_functions_scalar_datetime::datetime::parse_date_with_auto;
-use databend_functions_scalar_datetime::datetime::parse_epoch_str;
-use databend_functions_scalar_datetime::datetime::parse_timestamp_tz_with_auto;
 
 use crate::InputCommonSettings;
