+++ src/query/formats/src/field_decoder/json_ast.rs
@@ -34,6 +34,10 @@ use databend_common_expression::types::decimal::DecimalSize;
 use databend_common_expression::types::nullable::NullableColumnBuilder;
 use databend_common_expression::types::number::Number;
 use databend_common_expression::types::string::StringColumnBuilder;
+use databend_common_expression::utils::auto_detect_datetime::int64_to_timestamp;
+use databend_common_expression::utils::auto_detect_datetime::parse_date_with_auto;
+use databend_common_expression::utils::auto_detect_datetime::parse_timestamp_tz_with_auto;
+use databend_common_expression::utils::auto_detect_datetime::parse_timestamp_with_auto;
 use databend_common_expression::with_decimal_type;
 use databend_common_expression::with_number_mapped_type;
 use databend_common_io::HybridBitmap;
@@ -42,10 +46,6 @@ use databend_common_io::geography::geography_from_ewkt;
 use databend_common_io::geometry_from_ewkt;
 use databend_common_io::parse_bitmap;
 use databend_common_io::prelude::InputFormatSettings;
-use databend_functions_scalar_datetime::datetime::int64_to_timestamp;
-use databend_functions_scalar_datetime::datetime::parse_date_with_auto;
-use databend_functions_scalar_datetime::datetime::parse_timestamp_tz_with_auto;
-use databend_functions_scalar_datetime::datetime::parse_timestamp_with_auto;
 use jiff::tz::TimeZone;
 use lexical_core::FromLexical;
 use num_traits::NumCast;
