+++ src/query/functions/src/scalars/variant.rs
@@ -68,14 +68,14 @@ use databend_common_expression::types::timestamp::clamp_timestamp;
 use databend_common_expression::types::timestamp_tz::TimestampTzType;
 use databend_common_expression::types::variant::cast_scalar_to_variant;
 use databend_common_expression::types::variant::cast_scalars_to_variants;
+use databend_common_expression::utils::auto_detect_datetime::parse_date_with_auto;
+use databend_common_expression::utils::auto_detect_datetime::parse_timestamp_tz_with_auto;
+use databend_common_expression::utils::auto_detect_datetime::parse_timestamp_with_auto;
 use databend_common_expression::vectorize_with_builder_1_arg;
 use databend_common_expression::vectorize_with_builder_2_arg;
 use databend_common_expression::vectorize_with_builder_3_arg;
 use databend_common_expression::with_number_mapped_type;
 use databend_common_io::Interval;
-use databend_functions_scalar_datetime::datetime::parse_date_with_auto;
-use databend_functions_scalar_datetime::datetime::parse_timestamp_tz_with_auto;
-use databend_functions_scalar_datetime::datetime::parse_timestamp_with_auto;
 use jiff::Timestamp;
 use jiff::tz::TimeZone;
 use jsonb::OwnedJsonb;
