+++ src/query/functions/tests/it/scalars/datetime.rs
@@ -18,10 +18,10 @@ use std::str::FromStr;
 use databend_common_expression::FromData;
 use databend_common_expression::FunctionContext;
 use databend_common_expression::types::*;
-use databend_functions_scalar_datetime::datetime::auto_detect_date;
-use databend_functions_scalar_datetime::datetime::auto_detect_timestamp;
-use databend_functions_scalar_datetime::datetime::auto_detect_timestamp_tz;
-use databend_functions_scalar_datetime::datetime::parse_epoch_str;
+use databend_common_expression::utils::auto_detect_datetime::auto_detect_date;
+use databend_common_expression::utils::auto_detect_datetime::auto_detect_timestamp;
+use databend_common_expression::utils::auto_detect_datetime::auto_detect_timestamp_tz;
+use databend_common_expression::utils::auto_detect_datetime::parse_epoch_str;
 use goldenfile::Mint;
 use jiff::Timestamp;
 use jiff::tz::TimeZone;
